// src/latex/latex.interfaces.ts
export type SemiRectDirection = "top" | "bottom" | "left" | "right"
export type CornerDirection =
  | "topLeft"
  | "topRight"
  | "bottomLeft"
  | "bottomRight"
export type Lengths = "cm" | "mm" | "in" | "pt" | "em" | "ex" | "mu"
export type Colors = "white" | "gray" | "black" | "red" | "green" | "blue" | "cyan" | "magenta" | "yellow" | "orange"
export type Thickness = "ultra thin" | "very thin" | "thin" | "semithick" | "thick" | "very thick" | "ultra thick"